backend/requirements.txt
Flask[async]==2.3.2
Flask-Cors==3.0.10
openai==1.3.0
python-dotenv==1.0.0
requests==2.31.0
scikit-learn>=1.4.0
numpy>=1.26.0
